Support creating distillery miniwheel in any pyroinsta checkout
Context: We plan to create an fbpkg that contains pyroinsta's virtualenv so that we can test the server orchestrator.

This diff removes the hard coding of /opt/pyroinsta and starts parsing the command-line arguments in a rudimentary way to support building wherever a pyroinsta checkout may be.

Based on Facebook D26996466