Make Pyro statically linkable
Summary: This diff includes two changes which set up Pyro to be statically linkable in the future. The first is to sprinkle in `-fPIC` to the compilation of static libraries in order to make them linkable with other Buck targets down the line, and the second is to convert the extension library to be a static library so that it can be linked against by other targets in the future.

Based on Facebook D27126978