Change header word when forwarding objects
When forwarding objects just overwrite the header word instead of using a special layout id for forwarded objects. It is easy to recognize forwarded objects by the header no longer being tagged as a header word.

Remove concept of a minimum object size which is no longer needed with this change.

Based on Facebook D26655873