Add whole-archive linker flag when linking extension library to pyro executable
D27126978 caused some .so's for modules built by native_build.py in pyroinsta to stop linking properly to Pyro. The root cause was that we changed the extension library from being an object library, where all symbols are linked unconditionally, to a static library where unused symbols are pruned by the linker.

This diff brings that behavior back by ensuring that we have the `--whole-archive` flag set when linking `extension`.

Based on Facebook D27174223