Remove Runtime::newTuple
Tuples are immutable structures, so the use of the `RawTuple::atPut` method is strongly discouraged. And instead `MutableTuples` should be constructed, assigned to and turned into `Tuple`s with `becomeImmutable`.

This removes the `Runtime::newTuple` function to encourage this programming style. Note that `RawTuple::atPut` will stay around for now to support the `PyTuple_SetItem` C-API.

Note that `newMutableTuple` is not allowed to be used with a 0 value in order to enforce the uniqueness of the empty tuple as an optimization. Some of the calling code now has to accomodate for that.

Based on Facebook D27147340