Store cache index inline in bytecode
Remove originalArgs and leave the arg in its original position in the bytecode.
Store cache indices for all opcodes (except cached global opcodes) in the
inline cache index in the bytecode.

Instead of passing the cache index to every opcode handler, make opcodes that
need a cache index fetch it from the bytecode stream.

Based on Facebook D26494589