Just expand bytecode; do nothing
This is a stepping stone in the process of adding cache index slots to the
bytecode. This diff adds space for them and ensures that the rest of the
interpreter still works. It also unfortunately has to go through and multiply
the arguments of opcodes that take arguments with byte offsets or byte indices,
to account for the change in code unit size from 2 to 4. This can get removed
in later diffs that modify the compiler.

Based on Facebook D26374694