Build CPython with debug in Pyro debug build
Use the `CMAKE_BUILD_TYPE` flag to check
if we should build CPython in debug mode.

In addition to a typo (`CMAKE_BUILDTYPE`),
this flag was getting appended with lowercase letters.
Since CMake variables are case sensitive,
we have to convert this flag `TOUPPER` first.

Fix some tests that did not work in debug CPython
due to memory bugs in the tests.

Based on Facebook D27399276