Fix PyBuffer_Release to avoid calling custom releasebuffer functions for all byteslikes
Summary: We had implemented support for all buffer protocol objects in `PyObject_GetBuffer`. However, we didn't properly update the release function to check for all byteslikes instead of just bytes and bytearrays. Let's fix this and fix our tests to check that no error was raised.

Based on Facebook D27448991