[rfc] Hard-code a python3.8 symlink when building
Right now, we are unable to build a lot of tp2 packages in platform009 as they expect a python3.8 binary to exist. Let's lie for now, and symlink python3.8 -> python3.7 to unblock adding pyro 3.8 versions to these libraries.

We would, of course, remove this hack as soon as we bump the version to 3.8 proper.

Based on Facebook D27654157