Modify build.py to rsync dependencies from a buck target instead of reading from /opt/instagram
The basic problem: Our virtualenv building depended on the state of two repositories, instagram-server and FBCode. We had a couple python files we stub out which we rsync from /opt/instagram/virtualenv, which depends on the current instagram-server commit. This causes multiple problems:

1. To get a run consistent with someone else's, you need 2 repo hashes, and furthermore there's no guarantee that the Python artifacts we pull are consistent with the venv we built on FBCode.

2. This hardcodes /opt/instagram being present, which complicates building a virtualenv on Sandcastle (it can be done, but the other consistency/hermeticity issues apply).

This diff adds a target to build the link tree of the targets we want with Cinder, and rsync's them, making everything nice and self-contained in FBCode.

Based on Facebook D27829123