Remove squashfs compression from pyro.virtualenv fbpkg
Summary: The squashfs compression is something we took from Instagram's server FBPkg. This is not what Instagram uses for the virtualenv FBPkg, and it errors when attempting to sync the venv-pyro38 directory due to symlinks to /mnt/third-party2 that we have for radix. This diff removes the squashfs bit to default to fbpkg's regular compression.

Based on Facebook D27861770