Address bug in handling of stop = -1 case in itertools.islice
Summary: The islice implementation uses a sentinel -1 value for an unspecified stop, instead of None. This meant that we could compare all indices to stop, which, when unspecified, compares less than every index. This caused any iteration with islice of stop = None to increment through the indices one by one. Let's fix this, and add a test.

Based on Facebook D27856703