Implement `signal.alarm`
Pretty straightforward. As in the CPython implementation, we basically defer to the syscall.

I chose to write the unit test in Python mostly due to convenience. `alarm` does not fail even when called with negative numbers, so let's test that (probably a bug in CPython, but shrug).

Based on Facebook D27913625