JIT: Support binary, compare operations on smallint
Add only the binary/compare opcodes that are implemented in the
assembly interpreter. The JIT cannot yet call an opcode handler written in C++.

This also introduces emitNextOpcodeFallthrough. This is meant to differentiate in the JIT
between two cases:

* Ending the current opcode by flowing immediately to the next
* Ending the current opcode by jumping over some other (often slow-path)
  code to the next

A simpler thing would be to always jump, but sometimes that leads to jumps like
the following:

    jmp A:
    A:

This is likely not a problem in practice since it's always an icache/branch
prediction hit, so we could just avoid this if we want.  But it makes the
compiled code smaller.

This diff also adds an array of labels mapping opcode indices (NOT PCs) to
compiled code locations. This is needed for emitNextOpcodeFallthrough.

Also add UNARY_NOT.

Based on Facebook D27754481