Add simple JIT for simple functions
Support some opcodes. LOAD/STORE_FAST... are trivial except that for
the generic handler it requires loading the oparg into kOpargReg. This can be
optimized for the JIT by directly substituting the known-at-compile-time
offset, but that is left for later to avoid too much code duplication.

Based on Facebook D27748707