Call `__set_name__` on attributes when constructing a type
We were missing this part of a type's initialization that's done in CPython. Cached properties do not work without `__set_name__` being called, which is the concrete problem this fixes.

Note that `_type_new` gets called during bootstrapping, which means that method lookups might hit types which don't have an MRO defined yet. This meant that we need some way of looking up methods in the presence of MRO = None. We solve this by checking the mro() of the type before looking up the method.

Based on Facebook D28098256