Fix `_mod_convert_number_int/index` to properly handle int subclasses
Summary: Our previous implementation of `_mod_convert_number_int` blindly returned an object if it was an instance of an int without properly converting it to an int representation. Constrain the intrinsic to only operate on bools and ints, and properly convert bools to ints before pushing them back on the stack.

Based on Facebook D28113347