Modify asyncio.tasks to use `_step` instead of `__step` as the primitive for running tasks
Little bit of context here is warranted. In Python 3.6, asyncio.Task had an internal `_step` function. The libfb RequestContextAwareTasks (the task class that Instagram uses) overrides the step implementation to maintain the global RequestContext state properly. In Python 3.7 `_step` got renamed to `__step`, which is bad for us: `__step` cannot be overridden due to the mangling that happens, so we were no longer properly maintaining request context state.

In D23974146 we tried to re-introduce `_step` as an accessible variable, but this doesn't accomplish what we want because the real reason you want access to that function is to be able to override it. Vlad also fixed this for Cinder using this exact approach in  D18605205.

What this diff does is rename `__step` to `_step` to preserve the RequestContextAwareTask behavior. Short of implementing a custom ContextAwareTask as Instagram did (a custom extension), this is the best we can do for now.

Based on Facebook D28152515