Add a step to build_fbcode.py to precompile bytecode at the end of the build
Summary: Add a bytecode compilation step at the end of the fbcode / miniwheel build. This is isn't stricly necessary since python compiles the files to bytecode as well as they are imported. However that happens sequentially while this step can be performed in parallel. Resuling in a 2 minutes longer miniwheel build but typically more than 10 minutes saved compared to python compiling the files as they come in.

Based on Facebook D28228342