Set kIsFixedAttributeBase on builtin types
With multiple inheritance from a mix of extension types and managed types computeFixedAttributeBase should pick the object with fixed attributes that is highest in the inheritance chain.
Using the test below as an example, ExtensionBase and object are both valid fixedAttributeBase's the inheritance chain looks like this:
```
                   | Base -> ExtensionBaseClass -> object
SubClass:
                   | SimpleManagedBaseClass -> object
```

In this example ExtensionBaseClass should be selected as the appropriate base.

We know check `type.isBuiltin` before assuming that the fixed attribute base for a type is builtinBase.

Based on Facebook D28274036