Do not re-write LOAD_FAST/STORE_FAST if EXTENDED_ARG may be used.
Summary: When re-writing LOAD_FAST/STORE_FAST to LOAD_FAST_REVERSED/STORE_FAST_REVERSED we calculate the updated arg based on the total number of locals in the function. If the number of total locals exceeds 255 the updated arg no longer fits in the 8 bit arg.

Based on Facebook D28367418