freeExtensionModules after cleaning up instances which may reference module state
On finalization some dealloc functions attempt to access the module state, this causes a segfault if the module state is cleared prior to the finalizer being called.
This diff does not guarantee the module state will exist but changes the finalization order to not free the modules until after the native instances have been cleaned up.

Based on Facebook D28425297