Implement PyCFunction_GET_SELF
The only difference between `PyCFunction_GET_SELF` and `PyCFunction_GetSelf` in CPython is that `PyCFunction_GetSelf` first checks to make sure that the argument is actually a `PyCFunction` before accessing its internals. However, since we need to switch on its type in order to get the appopriate `Function` out of it, we can't take the shortcut, so our implementation can just be calling `PyCFunction_GetSelf`.

While adding a test, I noticed that our `GetSelf` test never actually called `GetSelf`, so I changed the test to call the correct function.

Based on Facebook D28572258