Add return mode to entryAsm interface
Right now the only interface for `entryAsm` is that we have the
`function_entry_assignment` register map. To support function calls from the
JIT, we need to be able to set a return mode from the "caller" of `entryAsm`.
This is because the JIT doesn't have an interpreter loop, and the default mode
in `entryAsm` is to let the interpreter loop pick up the next instruction. We
need to instead signal that we have to return to a specific address in machine
code to resume execution.

Based on Facebook D27950127