JIT: Use return mode in CALL_FUNCTION
Given that we do most of our function calls as tailcalls, it's impossible to
use normal `call`/`ret`. The callee would never actually return. This diff adds
a new return mode, `JitReturn`, so that the callee knows what address to return
to in the JITed function.

Because actually pushing an address onto the Python stack would mess up
execution, we emulate `call`/`ret` by pushing and popping from the C++ stack.

Based on Facebook D28401153