JIT: Make C++ interpreter functions return DEOPT
There should be no mutable state in the JIT; all state changes should
be handled in deopt. This diff causes C++ cache invalidation functions to
return a new signal, `Continue::DEOPT`. This signal tells the JIT to deoptimize
the current function and try once more in the interpreter.

Based on Facebook D27982583