Compile list/set/dict comprehensions inline
This change code generation for list-comprehensions, set-comprehensions and dict-comprehensions from generating an ad-hoc inner function to instead produce the loop in-place.

The trick to get this working is to rename the iteration target (`generator.target`) to something unique to the expression. Here we prefix the name with `_gen$`. We then rename all of the generator elements/values (`generator.elt` and `generator.val` for dict-comprehensions).

This will have slight effects on the semantics in that new `_gen$` names will appear in the `locals()` lists and in module/type scopes someone could now notice `LOAD_NAME`/`STORE_NAME` happening rather than locals executing in an inner function, and there could be a name clash if someone inserts `$` names manually into the `globals()` dict in module/global namespace...

Based on Facebook D28581179