Remove testing::newEmptyCode; stop use Code::setCode in tests
The code object is/should be immutable. It should be possible for `Runtime::newCode` to assume that all fields have their final values.

This fixes a common anti-pattern where tests use `testing::newEmptyCode()` and changed the fields after the fact meaning the `Runtime::newCode` constructor would not see the final state of the code object. Need to clean this up for upcoming changes.

Added new helpers to facilitate this:
* `testing::newCodeWithBytes`
* `testing::newCodeWithBytesConsts`
* `testing::newCodeWithBytesConstsNames`
* `testing::newCodeWithBytesConstsLocals`

Based on Facebook D28764828