Remove unnecessary work from tuple C-API
`PyTuple_SET_ITEM`, `PyTuple_GetItem`, and `PyTuple_Size` are all rather hot functions in the `--inbox` endpoint.

Many of the calls to `PyTuple_GetItem` and `PyTuple_Size` are likely instances of `PyTuple_GET_ITEM` and `PyTuple_GET_SIZE`, which we can specialize to eliminate type checking.

Fix error behavior of `PyTuple_GetItem`, and fix a test of that behavior.

Based on Facebook D28806375