Return -1 from set_wakeup_fd instead of unimplemented
Again thanks to @sinancepel for actually doing the work :) D28761360

asyncio.loop.add_signal_handler uses set_wakeup_fd to verify that signals are only being added on the main thread.
https://www.internalfb.com/code/fbsource/[992ca96ed5298fd2e3f25fdda9716daba6ad68dd]/fbcode/pyro/third-party/cpython/Lib/asyncio/unix_events.py?lines=90-98

Since pyro does not allow for threads this check is not necessary, but is raising an Exception due to unimplemented.

https://docs.python.org/3/library/signal.html#signal.set_wakeup_fd
"The old wakeup fd is returned (or -1 if file descriptor wakeup was not enabled). If fd is -1, file descriptor wakeup is disabled."

Returning -1 signifies that file descriptor wakeup is not enabled without raising an unimplemented.

Based on Facebook D28883852